requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
aiohttp==3.10.5
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
openai==1.51.0
google-generativeai==0.8.3
